Update README.md
The FIS Alpine Skiing Championship is a well-loved sports tournament for athletes in skiing to compete with each other in the spirit of sports. It had brought great joy for athletes, coaches, and people all over the world.
In the past, the storing of data of each tournament season is done entirely on paper. However, as society has progressed, this method of storing data is no longer viable. The risks that come with physical text databases are simply unacceptable for such a large worldwide sports tournament.
In this assignment, we chose to develop an entirely new, computer-backed database for FIS Alpine Skiing Championship. It makes storing information such as player data or team data much easier.
New data can be easily inserted or removed without the complicity of a text-based database. Also, the database system is well-designed so we believe a lot of new things about the championship can be learned by our team.